Restructured directory for the project. Test files should be prefixed with test_*
